Move VaultReorderingPosition to seperate file
