+++ Vault/Sources/VaultFeed/Storage/VaultStore.swift
@@ -44,19 +44,6 @@ public protocol VaultStoreWriter: Sendable {
     func delete(id: UUID) async throws
 }
 
-public enum VaultReorderingPosition {
-    /// Position the item relative to the given item ID, before it.
-    case before(UUID)
-    /// Position the item relative to the given item ID, after it.
-    case after(UUID)
-
-    public var id: UUID {
-        switch self {
-        case let .before(uuid), let .after(uuid): uuid
-        }
-    }
-}
-
 /// @mockable
 public protocol VaultStoreExporter: Sendable {
     func exportVault(userDescription: String) async throws -> VaultApplicationPayload
